requirements.txt
pylint
Flask
pytest
requests
python-dotenv